pyproject.toml
[project]
name = "llm-cockpit"
version = "0.1.0a1"
description = "Multi-user web cockpit for Ollama: dashboard + chat + admin, pip-installable in one process"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "Bloxperts / Chris Unger" }]
keywords = ["ollama", "llm", "fastapi", "dashboard", "chat"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Framework :: FastAPI",
  "Intended Audience :: System Administrators",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.32",
  "httpx>=0.27",
  "sqlalchemy>=2.0",
  "alembic>=1.13",
  "pydantic>=2.9",
  "pydantic-settings>=2.5",
  "bcrypt>=4.2",
  "python-jose[cryptography]>=3.3",
  "sse-starlette>=2.1",
  "pyyaml>=6.0",
  "tomli-w>=1.0",
  "jinja2>=3.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "pytest-asyncio>=0.24",
  "ruff>=0.6",
]

[project.scripts]
cockpit-admin = "cockpit.cli:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["cockpit*"]

[tool.setuptools.package-data]
"cockpit" = [
  "default_config/*.yaml",
  "default_config/*.j2",
  "default_config/*.md",
  "migrations/script.py.mako",
  "migrations/versions/*.py",
  "migrations/env.py",
  "migrations/alembic.ini",
  "frontend_dist/**/*",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-q"

[tool.coverage.run]
branch = true
source = ["cockpit"]

[tool.ruff]
line-length = 110
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]
